chore: should not close server in CI

+++ scripts/teardown.setup.ts
@@ -5,7 +5,7 @@ const shouldCloseServer = process.env.SHOULD_CLOSE_SERVER === 'true';
 
 async function teardownSetup(config: FullConfig) {
   // 测试运行结束后，关闭服务
-  if (shouldCloseServer) {
+  if (shouldCloseServer && !process.env.CI) {
     deleteNocoBase();
   }
 }
